feat: improve poolPositionMover deploy logic
